requirements.txt
colorama==0.4.6
python-dotenv==1.0.1
Requests==2.31.0
